+++ processor/src/main/java/react4j/processor/ReactProcessor.java
@@ -1070,11 +1070,37 @@ public final class ReactProcessor
 
     if ( isArezComponent )
     {
+      ensureComputedMatchesExpectations( typeElement );
       final boolean runArezScheduler =
         hasAnyAutorunMethods( typeElement ) ||
         hasAnyKeepAliveComputedMethods( typeElement ) ||
         hasAnyDependencyMethods( typeElement );
       descriptor.setRunArezScheduler( runArezScheduler );
+
+      descriptor.setComputedMethods( getComputedMethods( typeElement ) );
+    }
+  }
+
+  private void ensureComputedMatchesExpectations( final @Nonnull TypeElement typeElement )
+  {
+    final TypeElement computedElement =
+      processingEnv.getElementUtils().getTypeElement( Constants.COMPUTED_ANNOTATION_CLASSNAME );
+    final Set<String> parameters = computedElement.getEnclosedElements()
+      .stream()
+      .map( e -> e.getSimpleName().toString() )
+      .collect( Collectors.toSet() );
+    if ( !( parameters.contains( "name" ) &&
+            parameters.contains( "priority" ) &&
+            parameters.contains( "keepAlive" ) &&
+            parameters.contains( "observeLowerPriorityDependencies" ) &&
+            4 == parameters.size() ) )
+    {
+      throw new ReactProcessorException( "The @" + Constants.COMPUTED_ANNOTATION_CLASSNAME + " annotation was " +
+                                         "expected to have the parameters name, priority. keepAlive and " +
+                                         "observeLowerPriorityDependencies but has " + parameters + ". The " +
+                                         "react4j annotation processor needs to be updated to handle " +
+                                         "the change in parameters.", typeElement );
+
     }
   }
 
@@ -1176,6 +1202,26 @@ public final class ReactProcessor
                                                  element,
                                                  annotationName,
                                                  parameterName ).getValue();
+  }
+
+  /**
+   * Return computed that have not had the priority parameter explicitly set.
+   */
+  @Nonnull
+  private List<MethodDescriptor> getComputedMethods( @Nonnull final TypeElement typeElement )
+  {
+    final DeclaredType type = (DeclaredType) typeElement.asType();
+    return getMethods( typeElement )
+      .stream()
+      .filter( method -> {
+        final AnnotationMirror mirror =
+          ProcessorUtil.findAnnotationByType( method, Constants.COMPUTED_ANNOTATION_CLASSNAME );
+        return null != mirror &&
+               mirror.getElementValues().keySet().stream()
+                 .noneMatch( v -> "priority".equals( v.getSimpleName().toString() ) );
+      } )
+      .map( m -> new MethodDescriptor( m, (ExecutableType) processingEnv.getTypeUtils().asMemberOf( type, m ) ) )
+      .collect( Collectors.toList() );
   }
 
   private void verifyNoDuplicateAnnotations( @Nonnull final ExecutableElement method )