+++ processor/src/test/resources/expected/com/example/arez/computed/React4j_BasicComputedComponent.java
@@ -1,4 +1,4 @@
-package com.example.arez;
+package com.example.arez.computed;
 
 import arez.annotations.ArezComponent;
 import arez.annotations.Computed;
@@ -16,34 +16,27 @@ import react4j.NativeAdapterComponent;
 import react4j.ReactConfig;
 
 @ArezComponent(
-    name = "KeepAliveComputedArezReactComponent",
-    deferSchedule = true
+    name = "BasicComputedComponent"
 )
 @Generated("react4j.processor.ReactProcessor")
-abstract class React4j_KeepAliveComputedArezReactComponent extends KeepAliveComputedArezReactComponent {
+abstract class React4j_BasicComputedComponent extends BasicComputedComponent {
   static final ComponentConstructorFunction TYPE = getConstructorFunction();
 
   @Nonnull
   private static ComponentConstructorFunction getConstructorFunction() {
     final ComponentConstructorFunction componentConstructor = NativeReactComponent::new;
     if ( ReactConfig.enableComponentNames() ) {
-      Js.asPropertyMap( componentConstructor ).set( "displayName", "KeepAliveComputedArezReactComponent" );
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "BasicComputedComponent" );
     }
     return componentConstructor;
   }
 
   @Override
-  protected final void triggerScheduler() {
-    getContext().triggerScheduler();
-  }
-
-  @Override
   @Computed(
-      priority = Priority.LOWEST,
-      keepAlive = true
+      priority = Priority.LOWEST
   )
-  int myAutorun() {
-    return super.myAutorun();
+  boolean isActive() {
+    return super.isActive();
   }
 
   @JsType(
@@ -63,15 +56,15 @@ abstract class React4j_KeepAliveComputedArezReactComponent extends KeepAliveComp
         @Nonnull JsPropertyMap<Object> arg1);
   }
 
-  private static final class NativeReactComponent extends NativeAdapterComponent<KeepAliveComputedArezReactComponent> implements Lifecycle {
+  private static final class NativeReactComponent extends NativeAdapterComponent<BasicComputedComponent> implements Lifecycle {
     @JsConstructor
     NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
       super( props );
     }
 
     @Override
-    protected KeepAliveComputedArezReactComponent createComponent() {
-      return new Arez_React4j_KeepAliveComputedArezReactComponent();
+    protected BasicComputedComponent createComponent() {
+      return new Arez_React4j_BasicComputedComponent();
     }
 
     @Override