+++ processor/src/test/resources/input/com/example/arez/computed/AlreadyPrioritizedComputedComponent.java
@@ -1,0 +1,26 @@
+package com.example.arez.computed;
+
+import arez.annotations.Computed;
+import arez.annotations.Priority;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent
+abstract class AlreadyPrioritizedComputedComponent
+  extends ReactArezComponent
+{
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @Computed( priority = Priority.HIGH )
+  boolean isActive()
+  {
+    return true;
+  }
+}