+++ processor/src/test/resources/input/com/example/arez/computed/AnnotatedComputedComponent.java
@@ -1,0 +1,27 @@
+package com.example.arez.computed;
+
+import arez.annotations.Computed;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent
+abstract class AnnotatedComputedComponent
+  extends ReactArezComponent
+{
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @Computed
+  @Nonnull
+  String isActive()
+  {
+    return "";
+  }
+}