+++ processor/src/test/resources/input/com/example/arez/computed/GenericsReturnComputedComponent.java
@@ -1,0 +1,26 @@
+package com.example.arez.computed;
+
+import arez.annotations.Computed;
+import java.util.function.Consumer;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent
+abstract class GenericsReturnComputedComponent
+  extends ReactArezComponent
+{
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+
+  @Computed
+  Consumer<String> isActive()
+  {
+    return null;
+  }
+}